refactor(model): split model definitions into focused files
- Separate identifiers, geometry, windows, workspaces, monitors, commands, events, parsing, and placement types
- Preserve existing DefiModel APIs and behavior
